CARDS-DEV: Refactored Deck to operate more like a stack. Simplified Shuffle and exported initializations of decks to deck factories to separate state (deck) from game logic, how a deck is composed.
Even though we really only use the StandardDeckFactory at this point, the Factory Pattern is still useful because we have decoupled the implementation of decks from their use. If we later need additional deck types, we can just construct them differently without having to extend the deck class. Usually, the object being provided by the factory is abstracted away as well, however given that different decks are defined more by the cards in them than by how the deck is used we will leave Deck as a single class for now.

In the future, we may want to implement other game types, such as Baccarat, which uses 6 52 card decks, Hold 'em, which uses a single 52 card deck, or Frog, which uses 36 cards